+++ base.js
@@ -33,6 +33,11 @@ module.exports = {
         'jest': true,
         'shared-node-browser': true,
       },
+      rules: {
+        // Reports modules without any exports, or with unused exports
+        // https://github.com/benmosher/eslint-plugin-import/blob/f63dd261809de6883b13b6b5b960e6d7f42a7813/docs/rules/no-unused-modules.md
+        'import/no-unused-modules': 'off'
+      }
     },
   ],
 };